dynparquet: move ValuesToBuffer from frostdb to dynparquet package
